pyproject.toml
[project]
name = "maia"
packages = [{ include = "maia", from = "src" }]
version = "0.1.0"
description = ""
authors = [
    {name = "Matheus bastos",email = "matheusjohonston2010@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg2 (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
